Fix SKU validation, remove TableLayout RestaurantId, and standardize prices to decimal

+++ Backend/Market.Application/Modules/Meal/Queries/GetMealsById/GetMealByIdDto.cs
@@ -5,7 +5,7 @@ public sealed class GetMealByIdDto
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Description { get; set; } = string.Empty;
-    public double BasePrice { get; set; }
+    public decimal BasePrice { get; set; }
     public bool IsAvailable { get; set; }
     public bool IsFeatured { get; set; }
     public bool StockManaged { get; set; }
